% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday}
\usage{
birthday(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Probability that 2 or more people in a random sample k have the same birthday.
}
\description{
Birthday
}
\examples{
birthday(10)
}
