% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebyshev.R
\name{chebyshev}
\alias{chebyshev}
\title{Chebyshev's Theorem}
\usage{
chebyshev(k)
}
\arguments{
\item{k}{A numeric value greater than 1, the number of standard deviations}
}
\value{
A  value between 0 and 1, the proportion given by Chebyshev's theorem.
}
\description{
Chebyshev's Theorem
}
\examples{
chebyshev(2)
}
