% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt-data.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT dataset}
\format{
A data frame with nrow(ddt) rows and 6 variables:
\describe{
  \item{DDT}{DDT concentration measurement (units as provided)}
  \item{LENGTH}{Fish length (mm)}
  \item{MILE}{River mile location where fish was sampled}
  \item{RIVER}{River name}
  \item{SPECIES}{Fish species}
  \item{WEIGHT}{Fish weight (g)}
}
}
\source{
Course materials for MATH4753
}
\usage{
ddt
}
\description{
Measurements of DDT concentrations and related variables.
}
\keyword{datasets}
