% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of bootstrap resamples}

\item{x}{numeric vector of data}

\item{fun}{statistic to compute on each sample}

\item{alpha}{significance level}

\item{cx}{character expansion size used for CI and point estimate}

\item{...}{additional arguments}
}
\value{
\item{ci}{Numeric vector of length 2 giving the bootstrap CI}
  \item{fun}{The statistic used}
  \item{x}{The original data vector}
}
\description{
myboot2
}
\examples{
\dontrun{myboot2(iter=10000, x=sam, fun="mean", alpha=0.05)}
}
