% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{a}{lower bound of distribution}

\item{b}{upper bound of distribution}
}
\value{
vector of sample means
}
\description{
mycltu
}
\examples{
mycltu(10,10000,0,5)
}
