% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{A string giving the name of the log-likelihood function to evaluate.}

\item{theta}{A numeric vector of parameter values over whihc the likliehood shoud be evaluated.}
}
\value{
A numeric value giving the maximum likelihood estimate.
}
\description{
mymaxlikg
}
\examples{
# mymaxlikg(lfun = "logbin2", theta = seq(0.01, 0.99, length = 100))
}
