% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{upper limit of shaded area}
}
\value{
A curve with a shaded region from -infinity to x=a and its area.
}
\description{
myncurve
}
\examples{
myncurve(mu=10, sigma=2, a=12)
}
