% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{}

\item{iter}{}

\item{time}{}
}
\value{
barplot
}
\description{
Creates iter barplots at a rate of 1 per time with a sample size n
}
\examples{
mysample(iter=1, n=1000, time=1)
}
