% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the number of seats on plane}

\item{gamma}{the probability of overbooking}

\item{p}{the probability that a passanger will show up for flight}
}
\value{
a list of nc, nd, N, p, gamma. a plot is alos returned
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N=400, gamma=0.02, p=0.95)}
}
