% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHO-data.R, R/ddt-data.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{Data from ddt}
\format{
## `who`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}

A data frame with nrow(ddt) rows and 6 variables:
\describe{
  \item{DDT}{DDT concentration measurement (units as provided)}
  \item{LENGTH}{Fish length (mm)}
  \item{MILE}{River mile location where fish was sampled}
  \item{RIVER}{River name}
  \item{SPECIES}{Fish species}
  \item{WEIGHT}{Fish weight (g)}
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>

Course materials for MATH4753
}
\usage{
ddt

ddt
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...

Measurements of DDT concentrations and related variables.
}
\keyword{datasets}
