% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt-data.R, R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT dataset}
\format{
A data frame with nrow(ddt) rows and 6 variables:
\describe{
  \item{DDT}{DDT concentration measurement (units as provided)}
  \item{LENGTH}{Fish length (mm)}
  \item{MILE}{River mile location where fish was sampled}
  \item{RIVER}{River name}
  \item{SPECIES}{Fish species}
  \item{WEIGHT}{Fish weight (g)}
}

## `who`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
Course materials for MATH4753

<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
ddt

ddt
}
\description{
Measurements of DDT concentrations and related variables.

A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
