% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsample.R
\name{testsample}
\alias{testsample}
\title{testsample}
\usage{
testsample(n, iter = 10, time = 0.5)
}
\arguments{
\item{time}{}
}
\value{
barplot
}
\description{
Creates iter barplots at a rate of 1 per time with a sample size n
}
\examples{
testsample(iter=1, n=1000, time=1)
}
